Fix amount overflow bug in create input

+++ Java/com_biblepaywallet_core_BRCoreTransactionInput.c
@@ -46,7 +46,7 @@ JNIEXPORT jlong JNICALL Java_com_biblepaywallet_core_BRCoreTransactionInput_crea
 
     input->txHash = UInt256Get((const void *) hashData);
     input->index = (uint32_t) index;
-    input->amount = (uint32_t) amount;
+    input->amount = (uint64_t) amount;
 
     // script
     input->script = NULL;
